requirements.txt
anthropic>=0.25.0
yfinance>=0.2.38
feedparser>=6.0.11
requests>=2.31.0
schedule>=1.2.1
python-dotenv>=1.0.0
pytz>=2024.1
